Updating undo redo shortcut handler

+++ app/client/src/components/layers/edit/Edit.vue
@@ -1368,6 +1368,7 @@ export default {
           break;
         case 7:
           editType = "addBldEntrance";
+          startCb = this.onBldEntranceInteractionStart;
           endCb = this.onBldEntranceInteractionEnd;
           break;
         default:
@@ -1505,6 +1506,7 @@ export default {
           prev_geom: this.beforeModifyFeature
         });
         this.featureRedoStack = [];
+        this.limitUndoRedotoFixedSize();
       }
 
       this.olEditCtrl.isInteractionOnProgress = false;
@@ -1563,6 +1565,18 @@ export default {
       this.olEditCtrl.popup.isVisible = true;
       //update cache
       this.updateFileInputFeatureCache();
+    },
+
+    /**
+     * Building Entrance interaction start.
+     */
+    onBldEntranceInteractionStart(e) {
+      if (e.features) {
+        this.beforeModifyFeature = e.features
+          .getArray()[0]
+          .getGeometry()
+          .clone();
+      }
     },
 
     /**
@@ -1733,11 +1747,20 @@ export default {
             }
             //Check if bldEntranceFeature is saved from UndoRedo functionality. Then ignore
             if (!featureUndoRedo) {
-              this.featureUndoStack.push({
-                type: "delete",
-                features: [bldEntranceFeature]
-              });
+              if (evt.type === "drawend") {
+                this.featureUndoStack.push({
+                  type: "delete",
+                  features: [bldEntranceFeature]
+                });
+              } else if (evt.type === "modifyend") {
+                this.featureUndoStack.push({
+                  type: "modifyBldEntr",
+                  features: [bldEntranceFeature],
+                  prev_geom: this.beforeModifyFeature
+                });
+              }
               this.featureRedoStack = [];
+              this.limitUndoRedotoFixedSize();
             } else {
               // Syncing building entrance feature across undo redo object stack
               undoFeatures.push(bldEntranceFeature);
@@ -1752,6 +1775,7 @@ export default {
                 bldEntranceFeature
               );
             }
+            this.olEditCtrl.source.changed();
           }
         });
       setTimeout(() => {
@@ -2182,6 +2206,7 @@ export default {
         this.featureRedoStack = [];
       }
       this.unDoRedoStatus = false;
+      this.limitUndoRedotoFixedSize();
     },
     /**
      * Methods used on popup cancel
@@ -2368,7 +2393,6 @@ export default {
     },
     syncDeletedBldEntrFeature(featureStack, oldFeature, newFeature) {
       // Syncing building entrance feature across undo redo object stack
-      let flag = false;
       if (featureStack) {
         for (let i = 0; i < featureStack.length; i++) {
           let f = featureStack[i];
@@ -2376,12 +2400,7 @@ export default {
             let subF = f.features[j];
             if (subF.getId() === oldFeature.getId()) {
               f.features[j] = newFeature;
-              flag = true;
-              break;
             }
-          }
-          if (flag === true) {
-            break;
           }
         }
       }
@@ -2483,6 +2502,15 @@ export default {
         this.olEditCtrl.featuresToCommit = [];
       }
     },
+    limitUndoRedotoFixedSize() {
+      // Maintain fixed size undo redo stack
+      if (this.featureUndoStack.length > this.undoRedoMaxLength) {
+        this.featureUndoStack.shift();
+      }
+      if (this.featureRedoStack.length > this.undoRedoMaxLength) {
+        this.featureRedoStack.shift();
+      }
+    },
     async unDoRedo(unre) {
       //This function will be called on click of undo/redo button
       this.olEditCtrl.featuresToCommit = [];
@@ -2527,6 +2555,25 @@ export default {
             undo_features.features[0].setProperties(newAttribute);
             this.olEditCtrl.featuresToCommit = undo_features.features;
             await this.olEditCtrl.transact(newAttribute);
+            this.unDoRedoStatus = false;
+            break;
+          }
+          case "modifyBldEntr": {
+            this.olEditCtrl.currentInteraction = "modify";
+            this.selectedLayer = this.editableLayers[0];
+            let newGeom = undo_features.prev_geom;
+            undo_features.prev_geom = undo_features.features[0].getGeometry();
+            undo_features.features[0].setGeometry(newGeom);
+            // this.olEditCtrl.featuresToCommit = undo_features.features;
+            let evt = {
+              type: "modifyend",
+              feature: undo_features.features[0]
+            };
+            await this.onBldEntranceInteractionEnd(
+              evt,
+              true,
+              this.undoFeatures
+            );
             this.unDoRedoStatus = false;
             break;
           }
@@ -2578,6 +2625,26 @@ export default {
             redo_features.features[0].setProperties(newAttribute);
             this.olEditCtrl.featuresToCommit = redo_features.features;
             await this.olEditCtrl.transact(newAttribute);
+            this.featureUndoStack.push(redo_features);
+            this.unDoRedoStatus = false;
+            break;
+          }
+          case "modifyBldEntr": {
+            this.olEditCtrl.currentInteraction = "modify";
+            this.selectedLayer = this.editableLayers[0];
+            let newGeom = redo_features.prev_geom;
+            redo_features.prev_geom = redo_features.features[0].getGeometry();
+            redo_features.features[0].setGeometry(newGeom);
+            // this.olEditCtrl.featuresToCommit = redo_features.features;
+            let evt = {
+              type: "modifyend",
+              feature: redo_features.features[0]
+            };
+            await this.onBldEntranceInteractionEnd(
+              evt,
+              true,
+              this.undoFeatures
+            );
             this.featureUndoStack.push(redo_features);
             this.unDoRedoStatus = false;
             break;
